requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
PyJWT==2.9.0
bcrypt==4.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
